Remove platform images from admin and refactor admin into components
